# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: PraticOS Integrations API
  description: |
    API de integrações do PraticOS para bot WhatsApp e integrações externas.

    ## Autenticação

    A API suporta três modos de autenticação:
    - **API Key**: Para integrações externas (Zapier, n8n, parceiros)
    - **Bot API Key**: Para o bot WhatsApp (Clawdbot)
    - **Bearer Token**: Para o app Flutter (Firebase ID Token)
  version: 1.0.0
  contact:
    name: PraticOS Support
    url: https://praticos.app
  license:
    name: Proprietary
    url: https://praticos.app/terms

servers:
  - url: https://southamerica-east1-praticos-app.cloudfunctions.net/api
    description: Production
  - url: http://localhost:5001/praticos-app/southamerica-east1/api
    description: Local Emulator

tags:
  - name: Auth
    description: Autenticação e tokens
  - name: Orders
    description: Ordens de serviço
  - name: Customers
    description: Clientes
  - name: Devices
    description: Dispositivos
  - name: Services
    description: Catálogo de serviços
  - name: Products
    description: Catálogo de produtos
  - name: Company
    description: Empresa e membros
  - name: Analytics
    description: Dashboard e relatórios
  - name: Bot Link
    description: Vinculação WhatsApp
  - name: Bot Invite
    description: Convites de colaborador
  - name: Bot Search
    description: Busca inteligente
  - name: Bot Summary
    description: Resumos formatados

components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key for external integrations
    ApiSecret:
      type: apiKey
      in: header
      name: X-API-Secret
      description: API Secret (required with API Key)
    BotApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Bot API Key for Clawdbot
    WhatsAppNumber:
      type: apiKey
      in: header
      name: X-WhatsApp-Number
      description: WhatsApp number in E.164 format
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID Token

  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Resource not found

    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean

    CustomerAggr:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string

    DeviceAggr:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        serial:
          type: string

    OrderSummary:
      type: object
      properties:
        id:
          type: string
        number:
          type: integer
        customer:
          $ref: '#/components/schemas/CustomerAggr'
        device:
          $ref: '#/components/schemas/DeviceAggr'
        status:
          type: string
          enum: [quote, approved, progress, done, canceled]
        total:
          type: number
        dueDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/OrderSummary'
        - type: object
          properties:
            services:
              type: array
              items:
                type: object
                properties:
                  service:
                    type: object
                  value:
                    type: number
                  description:
                    type: string
            products:
              type: array
              items:
                type: object
                properties:
                  product:
                    type: object
                  value:
                    type: number
                  quantity:
                    type: integer
            discount:
              type: number
            paidAmount:
              type: number
            remainingBalance:
              type: number

    Customer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        createdAt:
          type: string
          format: date-time

    Device:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        serial:
          type: string
        manufacturer:
          type: string
        category:
          type: string
        description:
          type: string

    Service:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        value:
          type: number
        photo:
          type: string

    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        value:
          type: number
        photo:
          type: string

    Company:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        segment:
          type: string
        country:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        logo:
          type: string

    CompanyMember:
      type: object
      properties:
        userId:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [owner, admin, supervisor, manager, consultant, technician]
        linkedChannels:
          type: array
          items:
            type: string

paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
                  version:
                    type: string

  /api/v1/auth/token:
    post:
      tags: [Auth]
      summary: Generate access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [apiKey, apiSecret]
              properties:
                apiKey:
                  type: string
                apiSecret:
                  type: string
      responses:
        '200':
          description: Token generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      expiresIn:
                        type: integer
                      companyId:
                        type: string

  /api/v1/orders:
    get:
      tags: [Orders]
      summary: List orders
      security:
        - ApiKey: []
          ApiSecret: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [quote, approved, progress, done, canceled]
        - name: customerId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Orders]
      summary: Create order
      security:
        - ApiKey: []
          ApiSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customerId]
              properties:
                customerId:
                  type: string
                deviceId:
                  type: string
                services:
                  type: array
                  items:
                    type: object
                    properties:
                      serviceId:
                        type: string
                      value:
                        type: number
                      description:
                        type: string
                products:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: string
                      quantity:
                        type: integer
                      value:
                        type: number
                dueDate:
                  type: string
                  format: date-time
                status:
                  type: string
                  enum: [quote]
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      number:
                        type: integer
                      status:
                        type: string

  /api/v1/orders/{id}:
    get:
      tags: [Orders]
      summary: Get order details
      security:
        - ApiKey: []
          ApiSecret: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/OrderDetail'

    patch:
      tags: [Orders]
      summary: Update order
      security:
        - ApiKey: []
          ApiSecret: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [quote, approved, progress, done, canceled]
                dueDate:
                  type: string
                  format: date-time
                assignedTo:
                  type: string
      responses:
        '200':
          description: Order updated

  /api/v1/orders/{id}/payments:
    post:
      tags: [Orders]
      summary: Add payment to order
      security:
        - ApiKey: []
          ApiSecret: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                type:
                  type: string
                  enum: [payment, discount]
                  default: payment
                description:
                  type: string
      responses:
        '200':
          description: Payment added
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      transactionId:
                        type: string
                      paidAmount:
                        type: number
                      remainingBalance:
                        type: number
                      isFullyPaid:
                        type: boolean

  /api/v1/customers:
    get:
      tags: [Customers]
      summary: List customers
      security:
        - ApiKey: []
          ApiSecret: []
      parameters:
        - name: phone
          in: query
          schema:
            type: string
        - name: email
          in: query
          schema:
            type: string
        - name: name
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of customers

    post:
      tags: [Customers]
      summary: Create customer
      security:
        - ApiKey: []
          ApiSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                phone:
                  type: string
                email:
                  type: string
                address:
                  type: string
      responses:
        '201':
          description: Customer created

  /api/v1/analytics/summary:
    get:
      tags: [Analytics]
      summary: Get analytics summary
      security:
        - ApiKey: []
          ApiSecret: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, year, custom]
            default: today
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Analytics summary

  /api/v1/analytics/pending:
    get:
      tags: [Analytics]
      summary: Get pending items
      security:
        - ApiKey: []
          ApiSecret: []
      responses:
        '200':
          description: Pending items

  /api/bot/link:
    post:
      tags: [Bot Link]
      summary: Link WhatsApp number
      security:
        - BotApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, whatsappNumber]
              properties:
                token:
                  type: string
                whatsappNumber:
                  type: string
      responses:
        '200':
          description: WhatsApp linked

    delete:
      tags: [Bot Link]
      summary: Unlink WhatsApp number
      security:
        - BotApiKey: []
        - WhatsAppNumber: []
      responses:
        '200':
          description: WhatsApp unlinked

  /api/bot/link/context:
    get:
      tags: [Bot Link]
      summary: Get user context
      security:
        - BotApiKey: []
        - WhatsAppNumber: []
      responses:
        '200':
          description: User context

  /api/bot/invite/create:
    post:
      tags: [Bot Invite]
      summary: Create collaborator invite
      security:
        - BotApiKey: []
        - WhatsAppNumber: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [collaboratorName, role]
              properties:
                collaboratorName:
                  type: string
                role:
                  type: string
                  enum: [technician, consultant, supervisor, manager]
      responses:
        '201':
          description: Invite created

  /api/bot/invite/accept:
    post:
      tags: [Bot Invite]
      summary: Accept invite
      security:
        - BotApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inviteCode, whatsappNumber]
              properties:
                inviteCode:
                  type: string
                whatsappNumber:
                  type: string
                name:
                  type: string
      responses:
        '200':
          description: Invite accepted

  /api/bot/customers/search:
    get:
      tags: [Bot Search]
      summary: Search customers
      security:
        - BotApiKey: []
        - WhatsAppNumber: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search results

  /api/bot/orders/quick:
    post:
      tags: [Bot Search]
      summary: Quick order creation
      security:
        - BotApiKey: []
        - WhatsAppNumber: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customerName, deviceName, problem]
              properties:
                customerName:
                  type: string
                customerPhone:
                  type: string
                deviceName:
                  type: string
                deviceSerial:
                  type: string
                problem:
                  type: string
                estimatedValue:
                  type: number
                dueDate:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Order created

  /api/bot/summary/today:
    get:
      tags: [Bot Summary]
      summary: Get today summary
      security:
        - BotApiKey: []
        - WhatsAppNumber: []
      responses:
        '200':
          description: Today summary with formatted message

  /api/bot/summary/pending:
    get:
      tags: [Bot Summary]
      summary: Get pending summary
      security:
        - BotApiKey: []
        - WhatsAppNumber: []
      responses:
        '200':
          description: Pending items with formatted message